// src/shared/types.ts
export interface Inscription {
  inscriptionId: string;
  inscriptionNumber: number;
  address: string;
  outputValue: number;
  preview: string;
  content: string;
  contentType: string;
  contentLength: number;
  timestamp: number;
  genesisTransaction: string;
  location: string;
  output: string;
  offset: number;
  contentBody: string;
}

export interface InscriptionMintedItem {
  title: string;
  desc: string;
  inscriptions: Inscription[];
}
// ...
export interface WalletHistory{
  transactionType: string;
  network: string;
  senderAddress: string;
  recevierAddress: string;
  transactionStatus: string;
  transactionHash: string;
  transactionTime: string;
  amount: string;
  networkFee: string;
  link: string;
}